+++ REvoDesign/common/magic_numbers.py
@@ -7,3 +7,5 @@ DEFAULT_CLUSTER_SCORE_MTX = 'PAM30'
 
 DEFAULT_PROFILE_TYPE_GROUP = ['PSSM', 'CSV', 'TSV', 'Pythia-ddG']  # default profile type
 DEFAULT_PROFILE_TYPE = 'PSSM'
+
+SIDECHAIN_SOLVER=['buildin', 'DLPacker']