+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- Sidechain solver:
+  - `buildin`: PyMOL mutagenesis with Dunbrack rotamer library.
+  - `DLPacker`: DLPacker.
+
 ## [1.3.1] - 2023-12-26
 
 ### Added